Update dependencies and fix sdk breaking changes.
