+++ .gitignore
@@ -1,23 +1,34 @@
+# Tooling
+**.classpath
+**.idea
+**.metadata
+**.project
+**.settings
+**eclipseBin
+
+# Build artifacts
+**.class
+**.iml
+**build
+**classes
+**libJar
+**out
+**target
+**/.gradle
+**javadoc
+**repository
+
 # Misc
-.DS_Store
+**.bak
+**.old
+**.log
+**.DS_Store
 
-# Gradle
-.gradle
-build
-
-# Project files
-*.iml
-.idea
-nodes.json*
-.classpath
-.key
-.nodes.json
-.nodes.json.bak
-.project
-.settings
-bin
-.gradle/
-.idea/
-*.log
-build/
-out/
+# Runtime files
+**.jks
+**.key
+**db
+**nodes.json
+**nodes*.zip
+**test.json
+**/*.log.*.zip
