+++ src/main/java/org/iot/dsa/dslink/jdbc/JDBCv2Helpers.java
@@ -1,29 +1,27 @@
-package org.iot.dsa.dslink.java.v2.jdbc;
+package org.iot.dsa.dslink.jdbc;
 
-import org.iot.dsa.node.DSList;
-
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.Driver;
+import java.sql.DriverManager;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.Enumeration;
 import java.util.Random;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import org.iot.dsa.node.DSList;
+import org.iot.dsa.node.DSNode;
 
 /**
  * Class for storing helper methods and variables.
+ *
  * @author James (Juris) Puchin
  * Created on 10/13/2017
  */
 class JDBCv2Helpers {
-    //////////////////////
-    //Global Static Data
-    ///////////////////////
-    private static DSList cashedDriversNames;
+
     private static final char[] ALPHA_CHARS;
     private static final Random RANDOM = new Random();
 
-    ////////////////////
-    //String Definitions
-    ////////////////////
     static final String DB_NAME = "DB Name";
     static final String DB_URL = "URL";
     static final String DB_USER = "User Name";
@@ -32,36 +30,46 @@ class JDBCv2Helpers {
     static final String ADD_DB = "Connect";
     static final String LAST_SUCCESS = "Last Success Con";
     static final String LAST_FAIL = "Last Fail Con";
-
     static final String EXT_ACCESS = "Allow Access";
     static final String CREATE_DB = "Create H2";
-
     static final String DRIVER_NAME = "Driver Name";
     static final String REGISTERED = "Registered";
     static final String ADD_DRIVER = "Add Driver";
-
     static final String REMOVE = "Disconnect";
-
     static final String QUERY = "Query";
     static final String UPDATE = "Update";
     static final String MAKE_NODES = "Nodes";
     static final String SHOW_TABLES = "Show Tables";
     static final String SHOW_TABLE = "Show Table";
-
     static final String EDIT = "Edit";
-
     static final String INTERVAL = "Query Interval";
     static final String STREAM_QUERY = "Streaming Query";
-
     static final String STATUS = "Connection Status";
 
-    ////////////////////////
-    //Helper Functions
-    ////////////////////////
+    private static DSList cashedDriversNames;
 
-    static void registerDriver(String driverClass) throws ClassNotFoundException {
-        Class.forName(driverClass);
-        cashedDriversNames = null;
+    static void cleanClose(ResultSet res, Statement stmt, Connection conn, DSNode node) {
+        if (res != null) {
+            try {
+                res.close();
+            } catch (SQLException e) {
+                node.warn(node.getPath(), e);
+            }
+        }
+        if (stmt != null) {
+            try {
+                stmt.close();
+            } catch (SQLException e) {
+                node.warn(node.getPath(), e);
+            }
+        }
+        if (conn != null) {
+            try {
+                conn.close();
+            } catch (SQLException e) {
+                node.warn(node.getPath(), e);
+            }
+        }
     }
 
     static DSList getRegisteredDrivers() {
@@ -80,28 +88,9 @@ class JDBCv2Helpers {
         return cashedDriversNames.copy();
     }
 
-    static void cleanClose(ResultSet res, Statement stmt, Connection conn, Logger log) {
-        if (res != null) {
-            try {
-                res.close();
-            } catch (SQLException e) {
-                log.log(Level.WARNING, "Error closing ResultTable", e);
-            }
-        }
-        if (stmt != null) {
-            try {
-                stmt.close();
-            } catch (SQLException e) {
-                log.log(Level.WARNING, "Error closing Statement", e);
-            }
-        }
-        if (conn != null) {
-            try {
-                conn.close();
-            } catch (SQLException e) {
-                log.log(Level.WARNING, "Error closing Connection", e);
-            }
-        }
+    static void registerDriver(String driverClass) throws ClassNotFoundException {
+        Class.forName(driverClass);
+        cashedDriversNames = null;
     }
 
     static {
